Break out QueryBuilder from ControllerFilters
The class is initialized with the default query (previously it received the model class then used `.all` as the default query).
