+++ lib/filterameter/controller_filters.rb
@@ -12,6 +12,7 @@ require 'filterameter/filter_factory'
 require 'filterameter/filter_declaration'
 require 'filterameter/log_subscriber'
 require 'filterameter/parameters_base'
+require 'filterameter/query_builder'
 
 module Filterameter
   # = Controller Filters
@@ -23,6 +24,7 @@ module Filterameter
     attr_writer :query_variable_name
 
     delegate :add_filter, to: :registry
+    delegate :build_query, to: :query_builder
 
     def initialize(controller_name, controller_path)
       @controller_name = controller_name
@@ -37,12 +39,8 @@ module Filterameter
       @query_variable_name ||= model_class.model_name.plural
     end
 
-    def build_query(filter_params, starting_query)
-      valid_filters(filter_params)
-        .tap { |parameters| convert_min_and_max_to_range(parameters) }
-        .reduce(starting_query || model_class.all) do |query, (name, value)|
-        registry.fetch(name).apply(query, value)
-      end
+    def query_builder
+      @query_builder ||= Filterameter::QueryBuilder.new(model_class.all, registry)
     end
 
     private
@@ -57,63 +55,6 @@ module Filterameter
     # lazy so that model_class can be optionally set
     def registry
       @registry ||= Filterameter::FilterRegistry.new(Filterameter::FilterFactory.new(model_class))
-    end
-
-    def valid_filters(filter_params)
-      remove_invalid_values(
-        remove_undeclared_filters(filter_params)
-      )
-    end
-
-    # if both min and max are present in the query parameters, replace with range
-    def convert_min_and_max_to_range(parameters)
-      registry.ranges.each do |attribute_name, min_max_names|
-        next unless min_max_names.values.all? { |min_max_name| parameters[min_max_name].present? }
-
-        parameters[attribute_name] = Range.new(parameters.delete(min_max_names[:min]),
-                                               parameters.delete(min_max_names[:max]))
-      end
-    end
-
-    def remove_undeclared_filters(filter_params)
-      filter_params.slice(*declared_parameter_names).tap do |declared_parameters|
-        handle_undeclared_parameters(filter_params) if declared_parameters.size != filter_params.size
-      end
-    end
-
-    def handle_undeclared_parameters(filter_params)
-      action = Filterameter.configuration.action_on_undeclared_parameters
-      return unless action
-
-      undeclared_parameter_names = filter_params.keys - declared_parameter_names
-      case action
-      when :log
-        ActiveSupport::Notifications.instrument('undeclared_parameters.filterameter', keys: undeclared_parameter_names)
-      when :raise
-        raise Filterameter::Exceptions::UndeclaredParameterError, undeclared_parameter_names
-      end
-    end
-
-    def remove_invalid_values(filter_params)
-      validator = validator_class.new(filter_params)
-      return filter_params if validator.valid?
-
-      case Filterameter.configuration.action_on_validation_failure
-      when :log
-        ActiveSupport::Notifications.instrument('validation_failure.filterameter', errors: validator.errors)
-      when :raise
-        raise Filterameter::Exceptions::ValidationError, validator.errors
-      end
-
-      filter_params.except(*validator.errors.keys.map(&:to_s))
-    end
-
-    def declared_parameter_names
-      registry.filter_names
-    end
-
-    def validator_class
-      @validator_class ||= Filterameter::ParametersBase.build_sub_class(registry.filter_declarations)
     end
   end
 end
